refactor(repository): migrate base repository logic to repository factory and update related imports
