+++ app/.server/repositories/notes-repository.ts
@@ -1,4 +1,4 @@
-import { BaseRepository } from "../base-repository";
+import { createRepository } from "../repository-factory";
 
 export type Note = {
   id: string;
@@ -9,7 +9,4 @@ export type NoteCreate = Omit<Note, "id">;
 export type NoteUpdate = Partial<Note>;
 export type NoteDelete = Pick<Note, "id">;
 
-// Abaixo, a classe BaseRepository é instanciada com o nome da coleção
-// e o tipo de dado que ela irá manipular.
-// Isso deve ser feito pra cada coleção que você deseja consumir ou manipular.
-export const NotesRepository = new BaseRepository<Note>("notes");
+export const NotesRepository = createRepository<Note>("notes");
